+++ zeno/zeno/nodes/prim/PrimitiveConnection.cpp
@@ -127,4 +127,33 @@ ZENDEFNODE(PrimitiveClearConnect,
     "primitive",
     }});
 
+struct PrimitiveLineSimpleLink : zeno::INode {
+    virtual void apply() override {
+        auto prim = get_input<zeno::PrimitiveObject>("prim");
+
+        prim->lines.clear();
+        intptr_t n = prim->verts.size();
+        for (intptr_t i = 1; i < n; i++) {
+            prim->lines.emplace_back(i - 1, i);
+        }
+        prim->lines.update();
+        set_output("prim", std::move(prim));
+    }
+};
+
+
+ZENDEFNODE(PrimitiveLineSimpleLink, {
+    {
+    {"PrimitiveObject", "prim"},
+    },
+    {
+    {"PrimitiveObject", "prim"},
+    },
+    {
+    },
+    {"primitive"},
+});
+
+
+
 }
